Soft Exception Recovery
It's been found with high uptimes, communication with discord webhook for one reason or another may fail, soft locking a portion of the bot.

This update, allows the bot to catch, print, and ignore these errors. This prevents the bot from being locked up and allows the user to simply try again to resolve the issue.

Misc change: Added a config template.